# contracts/openapi.yaml
openapi: 3.0.3
...
info:
  title: Bataille Navale — API
  version: 1.0.0
  description: |
    API d'une bataille navale à pouvoirs, jouable en solo contre l'IA ou en ligne à deux.

    **Serveur autoritaire.** Aucune réponse ne contient la position d'un navire adverse non
    découvert. Les grilles sont encodées en lignes de caractères (voir le schéma `BoardView`).

    **Identification.** `POST /games` et `POST /games/join` renvoient un `playerToken` opaque.
    Il doit accompagner toute requête suivante dans l'en-tête `X-Player-Token`, et la
    négociation SignalR dans la query string `?token=`. Ce n'est pas un mécanisme de sécurité
    mais d'identification de session.
servers:
  - url: https://localhost:7001
    description: Développement local
security:
  - PlayerToken: []
...
paths:
...
  /api/games:
    post:
      tags: [Games]
      summary: Crée une partie
      description: |
        L'appelant devient le joueur 1. En mode `SinglePlayer`, l'IA occupe immédiatement le
        second emplacement et la partie passe en `AwaitingDeployment`. En mode `PrivateOnline`,
        un `joinCode` est émis et la partie reste en `AwaitingOpponent`.
      operationId: createGame
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGameRequest' }
      responses:
        '201':
          description: Partie créée
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateGameResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/games/join:
    post:
      tags: [Games]
      summary: Rejoint une partie en attente
      description: Par `joinCode` (partie privée) ou par `gameId` (partie publique du salon).
      operationId: joinGame
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JoinGameRequest' }
      responses:
        '200':
          description: Partie rejointe
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JoinGameResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Partie déjà pleine ou démarrée
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ApiProblem' }
              examples:
                full:
                  value: { type: 'about:blank', title: Conflit, status: 409, detail: Cette partie a déjà deux joueurs., code: GAME_FULL }
  /api/games/open:
    get:
      tags: [Games]
      summary: Liste les parties publiques en attente
      operationId: listOpenGames
      security: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Salon
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OpenGame' }
  /api/games/{gameId}:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      tags: [Games]
      summary: Récupère l'état de la partie, filtré pour l'appelant
      description: |
        Point d'entrée de la reprise après rafraîchissement ou déconnexion. La réponse est
        construite pour le porteur du `X-Player-Token` : un autre joueur obtient une réponse
        différente pour la même partie.
      operationId: getGame
      responses:
        '200':
          description: État de la partie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/games/{gameId}/forfeit:
    parameters:
      - $ref: '#/components/parameters/GameId'
    post:
      tags: [Games]
      summary: Abandonne la partie
      operationId: forfeitGame
      responses:
        '200':
          description: Partie terminée par abandon
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameOver' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/games/{gameId}/events:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      tags: [Games]
      summary: Flux d'événements depuis un curseur
      description: |
        Sert au rattrapage après reconnexion et au rejeu d'une partie terminée. C'est aussi
        le repli fonctionnel si SignalR est indisponible sur le réseau de démonstration.
      operationId: getEvents
      parameters:
        - name: since
          in: query
          description: Dernier `sequence` déjà reçu. 0 pour tout l'historique.
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Événements filtrés pour l'appelant
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/GameEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────────────── Fleet ───────────────────────────────
...
  /api/games/{gameId}/spectate:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      tags: [Games]
      summary: Vue spectateur (E-09)
      description: |
        Les deux grilles, telles que chaque joueur les a révélées à l'autre — jamais la
        position d'un navire non coulé, y compris pour un tiers qui ne joue pas. Aucune
        action possible depuis cette vue ; pas de `X-Player-Token` requis.
      operationId: spectateGame
      security: []
      responses:
        '200':
          description: Vue spectateur
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpectatorView' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────────────── Catalog ───────────────────────────────
...
components:
  securitySchemes:
    PlayerToken:
      type: apiKey
      in: header
      name: X-Player-Token
      description: |
        Jeton opaque émis à la création ou à l'adhésion. Identifie la session du joueur et
        détermine quelle vue filtrée le serveur construit. Ce n'est pas une authentification.
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
    Unauthorized:
      description: En-tête X-Player-Token absent
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
    Forbidden:
      description: Le jeton ne correspond à aucun joueur de cette partie
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
    NotFound:
      description: Ressource inconnue
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiProblem' }
  schemas:
...
    GameMode:        { type: string, enum: [SinglePlayer, PrivateOnline, PublicOnline] }
    GameStatus:      { type: string, enum: [AwaitingOpponent, AwaitingDeployment, InProgress, Finished, Abandoned] }
    AiLevel:         { type: string, enum: [Random, HuntTarget, Probability] }
...
    ShipType:        { type: string, enum: [Carrier, Battleship, Cruiser, Submarine, Destroyer] }
...
    PlayerSlot:      { type: string, enum: [One, Two] }
...
    PowerSlotStatus: { type: string, enum: [Ready, Charging, Armed, OnCooldown, Exhausted] }
...
    PowerId:
      type: string
      enum:
        # Reconnaissance
        - Sonar
        - TacticalRadar
        - LineDrone
        - RadioIntercept
        - Satellite
        - WakeAnalysis
        # Offensif
        - TripleSalvo
        - CrossStrike
        - CarpetBombing
        - Torpedo
        - OrbitalStrike
        - PiercingMissile
        - NavalMine
        - HeavyBomb
        - TsarBomba
        # Défensif
        - Decoy
        - Shield
        - EmergencyRepair
        - EvasiveManeuver
        - Jamming
        - SmokeScreen
        # Utilitaire
        - DoubleTurn
        - Sabotage
        - Espionage
        - ReactorOverload

    # ── Base ──
    Coordinate:
      type: object
      required: [x, y]
      description: 0-based. x = colonne (0 → A), y = ligne (0 → 1).
      properties:
        x: { type: integer, minimum: 0, example: 4 }
        y: { type: integer, minimum: 0, example: 6 }
    BoardView:
      type: object
      required: [width, height, rows]
      description: |
        Grille encodée en lignes de caractères, une chaîne par ligne, un caractère par case.
        Légende commune : `.` inconnu · `o` manqué · `x` touché · `#` coulé · `~` révélé vide ·
        `?` révélé occupé · `f` masqué par un écran de fumée.
        Grille propre uniquement : `S` navire intact · `X` navire touché · `D` leurre ·
        `M` mine · `B` protégé par un bouclier.
      properties:
        width:  { type: integer, example: 10 }
        height: { type: integer, example: 10 }
        rows:
          type: array
          items: { type: string }
          example: ['..........', '....o.....', '..x#......', '..........']

    # ── Requêtes ──
    CreateGameRequest:
      type: object
      required: [playerName, mode, gridWidth, gridHeight, fleetPreset, powers, turnTimeoutSeconds]
      properties:
        playerName:  { type: string, minLength: 2, maxLength: 16, example: Amiral Lea }
        mode:        { $ref: '#/components/schemas/GameMode' }
        gridWidth:   { type: integer, minimum: 8, maximum: 16, default: 10 }
        gridHeight:  { type: integer, minimum: 8, maximum: 16, default: 10 }
        fleetPreset: { type: string, example: Classic }
        aiLevel:
          allOf: [{ $ref: '#/components/schemas/AiLevel' }]
          nullable: true
          description: Requis si mode = SinglePlayer, ignoré sinon.
        powers:
          type: array
          maxItems: 3
          description: >-
            Exactement 3 pouvoirs pour un loadout personnalisé ; une liste vide équipe Sonar,
            HeavyBomb et TsarBomba par défaut sur les deux joueurs (temporaire, en attendant la sélection de loadout
            complète, E-13).
          items: { $ref: '#/components/schemas/PowerId' }
        turnTimeoutSeconds:
          type: integer
          description: 0 = pas de limite de temps. Sinon 10 à 120.
          minimum: 0
          maximum: 120
          default: 30
        seed:
          type: integer
          nullable: true
          description: Graine de génération, pour rejouer une partie à l'identique.
    JoinGameRequest:
      type: object
      required: [playerName, powers]
      description: Fournir `joinCode` (partie privée) ou `gameId` (partie publique), pas les deux.
      properties:
        playerName: { type: string, minLength: 2, maxLength: 16 }
        joinCode:   { type: string, nullable: true, pattern: '^[A-HJ-NP-Z2-9]{6}$', example: K7M4XQ }
        gameId:     { type: string, format: uuid, nullable: true }
        powers:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/PowerId' }
...
    CreateGameResponse:
      type: object
      required: [gameId, playerToken, status]
      properties:
        gameId:      { type: string, format: uuid }
        playerToken: { type: string, format: uuid }
        joinCode:    { type: string, nullable: true, example: K7M4XQ }
        status:      { $ref: '#/components/schemas/GameStatus' }
    JoinGameResponse:
      type: object
      required: [gameId, playerToken, status, opponentName]
      properties:
        gameId:       { type: string, format: uuid }
        playerToken:  { type: string, format: uuid }
        status:       { $ref: '#/components/schemas/GameStatus' }
        opponentName: { type: string }
    OpenGame:
      type: object
      required: [gameId, hostName, gridWidth, gridHeight, fleetPreset, powersEnabled, turnTimeoutSeconds, createdAtUtc]
      properties:
        gameId:             { type: string, format: uuid }
        hostName:           { type: string }
        gridWidth:          { type: integer }
        gridHeight:         { type: integer }
        fleetPreset:        { type: string }
        powersEnabled:      { type: boolean }
        turnTimeoutSeconds: { type: integer }
        createdAtUtc:       { type: string, format: date-time }
    ShipState:
      type: object
      required: [id, type, size, hits, isSunk]
      properties:
        id:     { type: string, example: own-cruiser-1 }
        type:   { $ref: '#/components/schemas/ShipType' }
        size:   { type: integer }
        hits:   { type: integer }
        isSunk: { type: boolean }
        cells:
          type: array
          nullable: true
          description: |
            Renseigné pour sa propre flotte, ou pour un navire adverse déjà coulé.
            **Null dans tous les autres cas** — un navire adverse vivant ne révèle jamais ses cases.
          items: { $ref: '#/components/schemas/Coordinate' }
...
    PowerSlot:
      type: object
      required: [powerId, status, chargeRemaining, cooldownRemaining, usesLeft, canAffordNow]
      properties:
        powerId:           { $ref: '#/components/schemas/PowerId' }
        status:            { $ref: '#/components/schemas/PowerSlotStatus' }
        chargeRemaining:   { type: integer }
        cooldownRemaining: { type: integer }
        usesLeft:          { type: integer }
        canAffordNow:      { type: boolean }
    OpponentCharge:
      type: object
      required: [isCharging]
      description: |
        Ce que l'appelant sait de la charge adverse. Le pouvoir n'est identifié que s'il a été
        révélé — par Espionnage, ou par la révélation automatique de la Frappe orbitale.
      properties:
        isCharging:      { type: boolean }
        revealedPowerId: { allOf: [{ $ref: '#/components/schemas/PowerId' }], nullable: true }
        turnsRemaining:  { type: integer, nullable: true }
        revealedColumn:  { type: integer, nullable: true }
    SelfView:
      type: object
      required: [playerId, name, slot, energy, board, fleet, powers]
      properties:
        playerId: { type: string, format: uuid }
        name:     { type: string }
        slot:     { $ref: '#/components/schemas/PlayerSlot' }
        energy:   { type: integer }
        board:    { $ref: '#/components/schemas/BoardView' }
        fleet:
          type: array
          items: { $ref: '#/components/schemas/ShipState' }
        powers:
          type: array
          items: { $ref: '#/components/schemas/PowerSlot' }
    OpponentView:
      type: object
      required: [playerId, name, isConnected, isAi, energy, shipsRemaining, shipsTotal, targetBoard, sunkShips, charge, equippedPowers]
      properties:
        playerId:       { type: string, format: uuid }
        name:           { type: string }
        isConnected:    { type: boolean }
        isAi:           { type: boolean }
        energy:         { type: integer, description: Publique, pour permettre d'anticiper les pouvoirs. }
        shipsRemaining: { type: integer }
        shipsTotal:     { type: integer }
        targetBoard:
          allOf: [{ $ref: '#/components/schemas/BoardView' }]
          description: Grille adverse telle que l'appelant l'a découverte, jamais davantage.
        sunkShips:
          type: array
          items: { $ref: '#/components/schemas/ShipState' }
        charge: { $ref: '#/components/schemas/OpponentCharge' }
        equippedPowers:
          type: array
          items: { $ref: '#/components/schemas/PowerId' }
    GameState:
      type: object
      required: [gameId, mode, status, fleetPreset, turnNumber, self, opponent, recentEvents, eventCursor]
      properties:
        gameId:          { type: string, format: uuid }
        mode:            { $ref: '#/components/schemas/GameMode' }
        status:          { $ref: '#/components/schemas/GameStatus' }
        joinCode:        { type: string, nullable: true }
        fleetPreset:
          type: string
          description: |
            Indispensable au joueur qui a rejoint par code ou depuis le salon public : lui
            seul n'a pas choisi ce preset à la création.
          example: Classic
        turnNumber:      { type: integer }
        currentPlayerId: { type: string, format: uuid, nullable: true }
        turnDeadlineUtc: { type: string, format: date-time, nullable: true }
        winnerId:        { type: string, format: uuid, nullable: true }
        self:            { $ref: '#/components/schemas/SelfView' }
        opponent:        { $ref: '#/components/schemas/OpponentView' }
        recentEvents:
          type: array
          items: { $ref: '#/components/schemas/GameEvent' }
        eventCursor:
          type: integer
          description: Dernier `sequence` inclus. À passer à GET /events?since= après reconnexion.
    SpectatorPlayerView:
      type: object
      required: [playerId, name, slot, isConnected, isAi, shipsRemaining, shipsTotal, board, sunkShips]
      description: Ce que l'ADVERSAIRE de ce joueur a découvert de sa flotte — jamais sa propre vue.
      properties:
        playerId:       { type: string, format: uuid }
        name:           { type: string }
        slot:           { $ref: '#/components/schemas/PlayerSlot' }
        isConnected:    { type: boolean }
        isAi:           { type: boolean }
        shipsRemaining: { type: integer }
        shipsTotal:     { type: integer }
        board:          { $ref: '#/components/schemas/BoardView' }
        sunkShips:
          type: array
          items: { $ref: '#/components/schemas/ShipState' }
    SpectatorView:
      type: object
      required: [gameId, mode, status, turnNumber, player1, player2, recentEvents]
      properties:
        gameId:          { type: string, format: uuid }
        mode:            { $ref: '#/components/schemas/GameMode' }
        status:          { $ref: '#/components/schemas/GameStatus' }
        turnNumber:      { type: integer }
        currentPlayerId: { type: string, format: uuid, nullable: true }
        turnDeadlineUtc: { type: string, format: date-time, nullable: true }
        winnerId:        { type: string, format: uuid, nullable: true }
        player1:         { $ref: '#/components/schemas/SpectatorPlayerView' }
        player2:         { $ref: '#/components/schemas/SpectatorPlayerView' }
        recentEvents:
          type: array
          items: { $ref: '#/components/schemas/GameEvent' }
...
    GameEvent:
      type: object
      required: [sequence, atUtc, type, message]
      properties:
        sequence: { type: integer }
        atUtc:    { type: string, format: date-time }
        type:
          type: string
          enum: [ShotFired, ShipSunk, PowerActivated, PowerCharging, PowerAborted, PowerResolved,
                 EnergyChanged, TurnChanged, PlayerJoined, PlayerLeft, EmoteSent, GameOver]
        actorId:  { type: string, format: uuid, nullable: true }
        message:  { type: string, description: Texte prêt à afficher dans le journal. }
        data:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: Charge utile spécifique au type, déjà filtrée pour le destinataire.
    GameOver:
      type: object
      required: [gameId, winnerId, winnerName, reason, totalTurns, stats]
      properties:
        gameId:     { type: string, format: uuid }
        winnerId:   { type: string, format: uuid }
        winnerName: { type: string }
        reason:     { type: string, enum: [FleetDestroyed, Forfeit, Disconnected, Timeout] }
        totalTurns: { type: integer }
        stats:
          type: array
          items: { $ref: '#/components/schemas/PlayerStats' }
    PlayerStats:
      type: object
      required: [playerId, name, shotsFired, hits, accuracy, powersUsed, energySpent]
      properties:
        playerId:    { type: string, format: uuid }
        name:        { type: string }
        shotsFired:  { type: integer }
        hits:        { type: integer }
        accuracy:    { type: number, format: double, example: 0.41 }
        powersUsed:  { type: integer }
        energySpent: { type: integer }
...
    ApiProblem:
      type: object
      description: ProblemDetails (RFC 9457) enrichi d'un code métier stable.
      required: [title, status, detail, code]
      properties:
        type:   { type: string, example: 'about:blank' }
        title:  { type: string, example: Conflit }
        status: { type: integer, example: 409 }
        detail: { type: string, example: Ce n'est pas votre tour. }
        code:
          type: string
          description: |
            Le front réagit à ce code, jamais au texte. Valeurs : VALIDATION_FAILED,
            MISSING_TOKEN, NOT_A_PLAYER, GAME_NOT_FOUND, GAME_FULL, INVALID_JOIN_CODE,
            NOT_YOUR_TURN, GAME_NOT_IN_PROGRESS, CELL_ALREADY_TARGETED, OUT_OF_BOUNDS,
            OVERLAPPING_SHIPS, FLEET_INCOMPLETE, FLEET_ALREADY_PLACED, INSUFFICIENT_ENERGY,
            POWER_ON_COOLDOWN, POWER_NOT_EQUIPPED, POWER_EXHAUSTED, POWER_ALREADY_CHARGING,
            INVALID_TARGET, CARRIER_REQUIRED, SHIP_ALREADY_DAMAGED.
          example: NOT_YOUR_TURN
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: array
            items: { type: string }